send

+++ ups/src/main/java/org/example/domain/ResendACKsD.java
@@ -10,15 +10,15 @@ public class ResendACKsD {
     private Long ackId;
 
     @Column(name = "ack", nullable = false)
-    private Integer ack;
+    private Long ack;
     // Getters and setters
 
 
-    public Integer getAck() {
+    public Long getAck() {
         return ack;
     }
 
-    public void setAck(Integer ack) {
+    public void setAck(Long ack) {
         this.ack = ack;
     }
